% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{base64url_decoder}
\alias{base64url_decoder}
\title{Decodage base64url}
\usage{
base64url_decoder(x)
}
\arguments{
\item{x}{Chaine base64url.}
}
\value{
Un vecteur \code{raw}.
}
\description{
Decodage base64url
}
