% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empreinte.R
\name{empreinte_depuis_hex}
\alias{empreinte_depuis_hex}
\title{Empreinte depuis une chaine hexadecimale}
\usage{
empreinte_depuis_hex(x)
}
\arguments{
\item{x}{Chaine hexadecimale de 64 caracteres.}
}
\value{
Vecteur \code{raw} de 32 octets.
}
\description{
Empreinte depuis une chaine hexadecimale
}
