% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empreinte.R
\name{empreinte_hex}
\alias{empreinte_hex}
\title{Empreinte en hexadecimal}
\usage{
empreinte_hex(x)
}
\arguments{
\item{x}{Vecteur \code{raw}.}
}
\value{
Chaine hexadecimale minuscule.
}
\description{
Empreinte en hexadecimal
}
