% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{foret_creer}
\alias{foret_creer}
\title{Creation d'une foret}
\usage{
foret_creer(
  con,
  nom,
  regime,
  proprietaire = NULL,
  date_application_regime_forestier = NULL,
  surface_ha = NULL,
  id = uuid_v4()
)
}
\arguments{
\item{con}{Connexion DBI.}

\item{nom}{Nom de la foret.}

\item{regime}{L'un de \link{SOMMIER_REGIMES}.}

\item{proprietaire}{Proprietaire (facultatif).}

\item{date_application_regime_forestier}{Date d'application du regime
forestier - sans objet en foret privee.}

\item{surface_ha}{Surface en hectares (facultatif).}

\item{id}{UUID a affecter ; genere si absent.}
}
\value{
L'UUID de la foret creee.
}
\description{
Creation d'une foret
}
