% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{jws_verifier_detache}
\alias{jws_verifier_detache}
\title{Verification d'une signature JWS detachee}
\usage{
jws_verifier_detache(jws, charge, cle_publique)
}
\arguments{
\item{jws}{Jeton \code{en-tete..signature}.}

\item{charge}{Vecteur \code{raw} signe.}

\item{cle_publique}{Cle publique, lue par \code{\link[openssl:read_key]{openssl::read_pubkey()}}.}
}
\value{
\code{TRUE} si la signature est valide, \code{FALSE} sinon.
}
\description{
Verification d'une signature JWS detachee
}
\examples{
cle <- openssl::rsa_keygen(2048)
s <- signataire_cle(cle, claims = list(sub = "agent-01"))
charge <- openssl::rand_bytes(32)
jws <- jws_signer_detache(charge, s)
jws_verifier_detache(jws, charge, cle$pubkey)

}
