% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{NOM_FORET_DEMO}
\alias{NOM_FORET_DEMO}
\title{Nom de la foret du jeu de demonstration}
\format{
An object of class \code{character} of length 1.
}
\usage{
NOM_FORET_DEMO
}
\description{
Le nom porte la mention « jeu de demonstration » pour qu'aucune lecture de
la base ne puisse prendre ces ecritures pour des enregistrements reels.
}
\keyword{datasets}
