% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{signataire_cle}
\alias{signataire_cle}
\title{Signataire adosse a une cle privee}
\usage{
signataire_cle(cle, claims, kid = NULL)
}
\arguments{
\item{cle}{Cle privee lue par \code{\link[openssl:read_key]{openssl::read_key()}}.}

\item{claims}{Liste nommee des claims d'identite.}

\item{kid}{Identifiant de cle (facultatif).}
}
\value{
Un objet \code{sommier_signataire}.
}
\description{
Signataire adosse a une cle privee
}
\examples{
cle <- openssl::rsa_keygen(2048)
signataire_cle(cle, claims = list(sub = "agent-01", name = "Maire"))

}
