% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre6.R
\docType{data}
\name{SOMMIER_CONCURRENCES}
\alias{SOMMIER_CONCURRENCES}
\title{Niveaux de concurrence d'une plantation}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_CONCURRENCES
}
\description{
Niveaux de concurrence d'une plantation
}
