% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre7.R
\docType{data}
\name{SOMMIER_DISPOSITIFS_FISCAUX}
\alias{SOMMIER_DISPOSITIFS_FISCAUX}
\title{Dispositifs fiscaux de la foret privee}
\format{
An object of class \code{character} of length 6.
}
\usage{
SOMMIER_DISPOSITIFS_FISCAUX
}
\description{
Rattachement facultatif d'une ecriture a un dispositif fiscal, prevu par le
brief pour la foret privee.
}
\keyword{datasets}
