% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre9.R
\docType{data}
\name{SOMMIER_ETATS_SANITAIRES}
\alias{SOMMIER_ETATS_SANITAIRES}
\title{Etats sanitaires releves sur un sujet remarquable}
\format{
An object of class \code{character} of length 5.
}
\usage{
SOMMIER_ETATS_SANITAIRES
}
\description{
Etats sanitaires releves sur un sujet remarquable
}
\keyword{datasets}
