% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre6.R
\docType{data}
\name{SOMMIER_EXECUTIONS_TRAVAUX}
\alias{SOMMIER_EXECUTIONS_TRAVAUX}
\title{Modes d'execution des travaux}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_EXECUTIONS_TRAVAUX
}
\description{
Modes d'execution des travaux
}
