% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\docType{data}
\name{SOMMIER_FORMATS_QUARTO}
\alias{SOMMIER_FORMATS_QUARTO}
\title{Formats de rendu Quarto}
\format{
An object of class \code{character} of length 2.
}
\usage{
SOMMIER_FORMATS_QUARTO
}
\description{
Formats de rendu Quarto
}
\keyword{datasets}
