% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travaux_terrain.R
\name{sommier_installer_placette}
\alias{sommier_installer_placette}
\title{Installer une placette de suivi}
\usage{
sommier_installer_placette(
  con,
  travaux_id,
  code_placette,
  geometrie,
  auteur,
  date_evenement = Sys.Date(),
  id = uuid_v4(),
  ...
)
}
\arguments{
\item{con}{Connexion DBI.}

\item{travaux_id}{UUID de l'intervention suivie.}

\item{code_placette}{Code lisible ("P35-03").}

\item{geometrie}{Le centre, un point en WGS84 (\code{\link[=geom_point]{geom_point()}}).}

\item{auteur}{Compte qui ecrit.}

\item{date_evenement}{Date d'installation.}

\item{id}{UUID de l'entree ; un import de terrain y met celui du releve.}

\item{...}{Champs de \code{\link[=registre6_placette]{registre6_placette()}} : \code{rayon_m},
\code{materialisation}, \code{precision_m}, \code{source_gnss}, \code{observations}.}
}
\value{
Invisiblement, l'entree chainee (liste d'une entree).
}
\description{
Inscrit au registre 6 une placette permanente qui suivra une plantation ou
une regeneration : son centre, son rayon, et les travaux qu'elle suit.
}
\details{
Les travaux suivis doivent etre une intervention du registre 6 de la meme
foret, dans la meme unite de gestion : une placette de l'UG 35 ne suit pas
une plantation de l'UG 12. Le code d'une placette est unique dans son
unite ; une placette deplacee se rectifie, elle ne se reinstalle pas sous
le meme code.
}
\seealso{
\code{\link[=sommier_controler_placette]{sommier_controler_placette()}}, \code{\link[=registre6_placette]{registre6_placette()}}
}
