% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre6.R
\docType{data}
\name{SOMMIER_PREVUS_TRAVAUX}
\alias{SOMMIER_PREVUS_TRAVAUX}
\title{Travaux prevus, reportes ou non prevus}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_PREVUS_TRAVAUX
}
\description{
Constate a la reception : \code{prevu} (au document de gestion), \code{reporte}
(prevu un autre exercice), \code{non_prevu}. Hors \code{prevu}, un motif est exige.
}
