% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{data}
\name{SOMMIER_REFERENTIELS}
\alias{SOMMIER_REFERENTIELS}
\title{Referentiels d'export de la gestion anterieure}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_REFERENTIELS
}
\description{
\itemize{
\item \code{psg} : section « gestion anterieure » du plan simple de gestion
(bloc 3 de l'arrete du 19 juillet 2012).
\item \code{amenagement} : bilan de l'amenagement precedent (partie 2 du document
d'amenagement ONF).
\item \code{ct88} : evaluation de fin de plan (etape 5 du CT88).
}
}
\keyword{datasets}
