% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registres.R
\docType{data}
\name{SOMMIER_REGIMES}
\alias{SOMMIER_REGIMES}
\title{Regimes de propriete forestiere}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_REGIMES
}
\description{
Regimes de propriete forestiere
}
\keyword{datasets}
