% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registres.R
\docType{data}
\name{SOMMIER_REGISTRES_OUVERTS}
\alias{SOMMIER_REGISTRES_OUVERTS}
\title{Registres ouverts a l'ecriture dans cette version}
\format{
An object of class \code{integer} of length 5.
}
\usage{
SOMMIER_REGISTRES_OUVERTS
}
\description{
Registres ouverts a l'ecriture dans cette version
}
\keyword{datasets}
