% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre4.R
\docType{data}
\name{SOMMIER_REVETEMENTS}
\alias{SOMMIER_REVETEMENTS}
\title{Natures de revetement de la voirie forestiere (imprime A50D)}
\format{
An object of class \code{character} of length 4.
}
\usage{
SOMMIER_REVETEMENTS
}
\description{
Natures de revetement de la voirie forestiere (imprime A50D)
}
\keyword{datasets}
