% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travaux_suivi.R
\docType{data}
\name{SOMMIER_SEUIL_REPRISE_PCT}
\alias{SOMMIER_SEUIL_REPRISE_PCT}
\title{Seuil de reprise d'une plantation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SOMMIER_SEUIL_REPRISE_PCT
}
\description{
Taux de reprise, en pourcentage, sous lequel une plantation appelle un
regarni. Le rapport le trace en trait sur la courbe de reprise par age.
}
