% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patrimoine.R
\docType{data}
\name{SOMMIER_SEUIL_TGB_CM}
\alias{SOMMIER_SEUIL_TGB_CM}
\title{Seuil de circonference des tres gros bois}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SOMMIER_SEUIL_TGB_CM
}
\description{
Circonference a 1,30 m au-dela de laquelle un arbre est compte comme tres
gros bois pour l'IBP. La valeur par defaut, 220 cm, correspond a environ
70 cm de diametre ; le seuil retenu varie selon la region et le referentiel,
d'ou le parametre de \code{\link[=sommier_elements_ibp]{sommier_elements_ibp()}}.
}
\keyword{datasets}
