% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horodatage.R
\docType{data}
\name{SOMMIER_STATUTS_TSA}
\alias{SOMMIER_STATUTS_TSA}
\title{Statuts PKIStatus (RFC 3161 section 2.4.2)}
\format{
An object of class \code{character} of length 6.
}
\usage{
SOMMIER_STATUTS_TSA
}
\description{
Statuts PKIStatus (RFC 3161 section 2.4.2)
}
\keyword{datasets}
