% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registres.R
\docType{data}
\name{SOMMIER_TYPES_COUPE}
\alias{SOMMIER_TYPES_COUPE}
\title{Types d'entree du registre 5 (coupes et recoltes)}
\format{
An object of class \code{character} of length 4.
}
\usage{
SOMMIER_TYPES_COUPE
}
\description{
\itemize{
\item \code{martelage} : volume martele en coupe reglee ou non reglee (imprime A50E).
\item \code{produit_accidentel} : chablis, bois sanitaires, volumes marteles hors
coupe prevue (imprime A50E, colonne "produits accidentels").
\item \code{bois_delivre} : delivrance, principalement l'affouage en foret communale
(imprime A50G, colonne "bois delivres") - bois martele, donc imputable.
\item \code{coupe_realisee} : releve de la coupe effectivement exploitee
(imprime A50F).
}
}
\keyword{datasets}
