% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre3.R
\docType{data}
\name{SOMMIER_TYPES_DROIT}
\alias{SOMMIER_TYPES_DROIT}
\title{Types d'entree du registre 3 (droits et concessions)}
\format{
An object of class \code{character} of length 6.
}
\usage{
SOMMIER_TYPES_DROIT
}
\description{
Types d'entree du registre 3 (droits et concessions)
}
\keyword{datasets}
