% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre8.R
\docType{data}
\name{SOMMIER_TYPES_EVENEMENT}
\alias{SOMMIER_TYPES_EVENEMENT}
\title{Types d'entree du registre 8}
\format{
An object of class \code{character} of length 4.
}
\usage{
SOMMIER_TYPES_EVENEMENT
}
\description{
Le registre 8 recouvre trois imprimes et un flux entrant :
\itemize{
\item \code{phenomene} : journal chronologique des phenomenes interessant la vie de
la foret (imprime A50K) - tempetes, incendies, crises sanitaires, gel.
\item \code{tableau_chasse} : prelevement cynegetique par saison et par espece
(imprime A50L).
\item \code{equilibre_gibier} : constat d'equilibre foret-gibier, obligatoire en
PSG depuis la LAAAF de 2014.
\item \code{detection} : phenomene propose par teledetection, en attente de
validation terrain (voir \code{\link[=sommier_importer_detections]{sommier_importer_detections()}}).
}
}
\keyword{datasets}
