% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre4.R
\docType{data}
\name{SOMMIER_TYPES_INFRASTRUCTURE}
\alias{SOMMIER_TYPES_INFRASTRUCTURE}
\title{Types d'entree du registre 4 (infrastructures)}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_TYPES_INFRASTRUCTURE
}
\description{
Types d'entree du registre 4 (infrastructures)
}
\keyword{datasets}
