% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registres.R
\docType{data}
\name{SOMMIER_TYPES_MARTELES}
\alias{SOMMIER_TYPES_MARTELES}
\title{Types d'entree imputables a la balance de possibilite}
\format{
An object of class \code{character} of length 3.
}
\usage{
SOMMIER_TYPES_MARTELES
}
\description{
La balance A50E confronte a la possibilite les \strong{volumes marteles}, soit
les coupes et les produits accidentels. \code{coupe_realisee} en est exclu a
dessein : la meme coupe est d'abord martelee (imprime A50E) puis exploitee
(imprime A50F), l'imputer deux fois doublerait le prelevement constate.
}
\seealso{
\code{\link[=sommier_balance_possibilite]{sommier_balance_possibilite()}}
}
\keyword{datasets}
