% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{SOMMIER_TYPES_PHOTO}
\alias{SOMMIER_TYPES_PHOTO}
\title{Types de photo admis au depot}
\usage{
SOMMIER_TYPES_PHOTO
}
\description{
Extensions reconnues et type de media associe. Le type est ecrit dans le
payload, l'extension sert au nom du fichier dans le depot.
}
