% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre9.R
\docType{data}
\name{SOMMIER_TYPES_REMARQUABLE}
\alias{SOMMIER_TYPES_REMARQUABLE}
\title{Types de fiche du patrimoine remarquable (serie A50 r/*)}
\format{
An object of class \code{character} of length 5.
}
\usage{
SOMMIER_TYPES_REMARQUABLE
}
\description{
Les six imprimes de la serie : arbre remarquable (r/a), peuplement
remarquable (r/p), vestige ou element culturel (r/c), espece protegee
(r/e et r/s), habitat remarquable (r/h).
}
\keyword{datasets}
