% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre1.R
\docType{data}
\name{SOMMIER_TYPES_VALIDATION}
\alias{SOMMIER_TYPES_VALIDATION}
\title{Types d'acte du registre 1}
\format{
An object of class \code{character} of length 7.
}
\usage{
SOMMIER_TYPES_VALIDATION
}
\description{
\itemize{
\item \code{visa_annuel} : le visa du responsable du niveau de gestion (imprime
A10), obligatoire et annuel.
\item \code{visa_direction} : le visa du niveau de direction, second cadre de
l'imprime A10.
\item \code{arrete} : arrete ministeriel ou prefectoral d'amenagement.
\item \code{deliberation} : deliberation du conseil municipal (foret communale).
\item \code{agrement} : agrement du PSG par le CRPF (foret privee).
\item \code{avenant} : avenant au document de gestion.
\item \code{engagement_fiscal} : engagement pris au titre d'un dispositif fiscal
(DEFI, Monichon, IFI).
}
}
\keyword{datasets}
