% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registre4.R
\docType{data}
\name{SOMMIER_USAGES_VOIRIE}
\alias{SOMMIER_USAGES_VOIRIE}
\title{Usages de la voirie forestiere (imprime A50D bis)}
\format{
An object of class \code{character} of length 4.
}
\usage{
SOMMIER_USAGES_VOIRIE
}
\description{
Usages de la voirie forestiere (imprime A50D bis)
}
\keyword{datasets}
