% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empreinte.R
\docType{data}
\name{SOMMIER_VERSION_CHAINE}
\alias{SOMMIER_VERSION_CHAINE}
\title{Version de l'algorithme de chainage}
\format{
An object of class \code{character} of length 1.
}
\usage{
SOMMIER_VERSION_CHAINE
}
\description{
Identifiant de l'algorithme d'empreinte, inclus dans les octets haches afin
que la chaine soit auto-descriptive : un verificateur lisant un export sait
quelle regle appliquer, et un changement d'algorithme ne peut pas passer
pour la meme chaine.
}
\keyword{datasets}
