% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifeste.R
\docType{data}
\name{SOMMIER_VERSION_MANIFESTE}
\alias{SOMMIER_VERSION_MANIFESTE}
\title{Version du format de manifeste}
\format{
An object of class \code{character} of length 1.
}
\usage{
SOMMIER_VERSION_MANIFESTE
}
\description{
Version du format de manifeste
}
\keyword{datasets}
