% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ug_fusionner}
\alias{ug_fusionner}
\title{Fusion d'unites de gestion}
\usage{
ug_fusionner(con, ug_uuids, numero_affichage, date_effet, serie_id = NULL)
}
\arguments{
\item{con}{Connexion DBI.}

\item{ug_uuids}{UUID des unites a fusionner (au moins deux).}

\item{numero_affichage}{Numero de l'unite resultante.}

\item{date_effet}{Date d'effet de la fusion.}

\item{serie_id}{Serie de l'unite resultante (facultatif : celle de la
premiere unite fusionnee par defaut).}
}
\value{
L'UUID de l'unite creee.
}
\description{
Clot les unites d'origine et cree l'unite resultante. La filiation a
plusieurs parents n'etant pas representable par la colonne \code{parent_uuid},
elle est portee par la table \code{ug_filiation}.
}
