% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ug_lire}
\alias{ug_lire}
\title{Lecture d'unites de gestion}
\usage{
ug_lire(con, ug_uuid = NULL, foret_id = NULL, actives_seulement = FALSE)
}
\arguments{
\item{con}{Connexion DBI.}

\item{ug_uuid}{UUID a lire ; toutes les unites de \code{foret_id} si absent.}

\item{foret_id}{UUID de la foret.}

\item{actives_seulement}{Ne rendre que les unites non cloturees.}
}
\value{
Un \code{data.frame}.
}
\description{
Lecture d'unites de gestion
}
