% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valider.R
\name{uuid_v4}
\alias{uuid_v4}
\title{Generation d'un UUID version 4}
\usage{
uuid_v4(n = 1L)
}
\arguments{
\item{n}{Nombre d'identifiants a produire.}
}
\value{
Un vecteur de caracteres.
}
\description{
Utilise le generateur d'aleas cryptographique d'openssl : les identifiants
d'unites de gestion ne doivent jamais entrer en collision ni etre
devinables, puisqu'ils sont stables a vie.
}
\examples{
uuid_v4()

}
