% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter.R
\name{exercice_definir}
\alias{exercice_definir}
\title{Fixation de la possibilite d'un exercice}
\usage{
exercice_definir(con, foret_id, annee, possibilite_m3_an)
}
\arguments{
\item{con}{Connexion DBI.}

\item{foret_id}{UUID de la foret.}

\item{annee}{Annee de l'exercice.}

\item{possibilite_m3_an}{Possibilite annuelle en metres cubes.}
}
\value{
Invisiblement, le nombre de lignes ecrites.
}
\description{
Fixation de la possibilite d'un exercice
}
