% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registres.R
\docType{data}
\name{SOMMIER_SCHEMA_VERSIONS}
\alias{SOMMIER_SCHEMA_VERSIONS}
\title{Versions de schema des payloads}
\format{
An object of class \code{character} of length 2.
}
\usage{
SOMMIER_SCHEMA_VERSIONS
}
\description{
Le payload est du JSONB versionne par type de registre (brief, section 4).
La version est hachee avec l'entree : une evolution de schema ne peut donc
pas se faire passer pour l'ancienne.
}
\keyword{datasets}
