% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horodatage.R
\name{tsa_horodater}
\alias{tsa_horodater}
\title{Obtention d'un jeton d'horodatage}
\usage{
tsa_horodater(empreinte, url, transport = tsa_transport_curl(), nonce = NULL)
}
\arguments{
\item{empreinte}{Vecteur \code{raw} de 32 octets a horodater.}

\item{url}{URL de l'autorite d'horodatage.}

\item{transport}{Fonction de transport, voir \code{\link[=tsa_transport_curl]{tsa_transport_curl()}}.}

\item{nonce}{Nonce a employer (facultatif).}
}
\value{
Un vecteur \code{raw} : le jeton d'horodatage.
}
\description{
Obtention d'un jeton d'horodatage
}
